fix formatting_test
