+++ gprimitives/src/lib.rs
@@ -23,7 +23,8 @@
 #![doc(html_logo_url = "https://docs.gear.rs/logo.svg")]
 #![doc(html_favicon_url = "https://gear-tech.io/favicons/favicon.ico")]
 
-use blake2::{digest::typenum::U32, Blake2b, Digest};
+pub mod utils;
+
 use core::{
     fmt,
     str::{self, FromStr},
@@ -37,23 +38,6 @@ use {
         TypeInfo,
     },
 };
-
-/// BLAKE2b-256 hasher state.
-type Blake2b256 = Blake2b<U32>;
-
-fn hash(data: &[u8]) -> [u8; 32] {
-    let mut ctx = Blake2b256::new();
-    ctx.update(data);
-    ctx.finalize().into()
-}
-
-fn hash_of_array<T: AsRef<[u8]>, const N: usize>(array: [T; N]) -> [u8; 32] {
-    let mut ctx = Blake2b256::new();
-    for data in array {
-        ctx.update(data);
-    }
-    ctx.finalize().into()
-}
 
 /// Message handle.
 ///
@@ -216,14 +200,14 @@ impl ActorId {
     /// Generates `ActorId` from given `CodeId` and `salt`.
     pub fn generate_from_user(code_id: CodeId, salt: &[u8]) -> Self {
         const SALT: &[u8] = b"program_from_user";
-        hash_of_array([SALT, code_id.as_ref(), salt]).into()
+        utils::hash_of_array([SALT, code_id.as_ref(), salt]).into()
     }
 
     /// Generates `ActorId` from given `CodeId`, `MessageId` and `salt`.
     pub fn generate_from_program(code_id: CodeId, salt: &[u8], message_id: MessageId) -> Self {
         //TODO: consider to move `message_id` to first param
         const SALT: &[u8] = b"program_from_wasm";
-        hash_of_array([SALT, message_id.as_ref(), code_id.as_ref(), salt]).into()
+        utils::hash_of_array([SALT, message_id.as_ref(), code_id.as_ref(), salt]).into()
     }
 }
 
@@ -272,7 +256,7 @@ impl MessageId {
         local_nonce: u128,
     ) -> MessageId {
         const SALT: &[u8] = b"external";
-        hash_of_array([
+        utils::hash_of_array([
             SALT,
             &block_number.to_le_bytes(),
             user_id.as_ref(),
@@ -284,7 +268,7 @@ impl MessageId {
     /// Generates `MessageId` for program outgoing message.
     pub fn generate_outgoing(origin_msg_id: MessageId, local_nonce: u32) -> MessageId {
         const SALT: &[u8] = b"outgoing";
-        hash_of_array([SALT, origin_msg_id.as_ref(), &local_nonce.to_le_bytes()]).into()
+        utils::hash_of_array([SALT, origin_msg_id.as_ref(), &local_nonce.to_le_bytes()]).into()
     }
 
     /// Generates `MessageId` for reply message depend on status code.
@@ -293,13 +277,13 @@ impl MessageId {
     /// BECAUSE AUTO-REPLY LOGIC DEPENDS ON PRE-DEFINED REPLY ID.
     pub fn generate_reply(origin_msg_id: MessageId) -> MessageId {
         const SALT: &[u8] = b"reply";
-        hash_of_array([SALT, origin_msg_id.as_ref()]).into()
+        utils::hash_of_array([SALT, origin_msg_id.as_ref()]).into()
     }
 
     /// Generates `MessageId` for signal message depend on status code.
     pub fn generate_signal(origin_msg_id: MessageId) -> MessageId {
         const SALT: &[u8] = b"signal";
-        hash_of_array([SALT, origin_msg_id.as_ref()]).into()
+        utils::hash_of_array([SALT, origin_msg_id.as_ref()]).into()
     }
 }
 
@@ -322,7 +306,7 @@ declare_primitive!(new zero into_bytes from_u64 from_h256 try_from_slice display
 impl CodeId {
     /// Generates `CodeId` from given code.
     pub fn generate(code: &[u8]) -> Self {
-        hash(code).into()
+        utils::hash(code).into()
     }
 }
 
@@ -358,6 +342,51 @@ impl ReservationId {
     /// Generates `ReservationId` from given message and nonce.
     pub fn generate(msg_id: MessageId, nonce: u64) -> Self {
         const SALT: &[u8] = b"reservation";
-        hash_of_array([SALT, msg_id.as_ref(), &nonce.to_le_bytes()]).into()
+        utils::hash_of_array([SALT, msg_id.as_ref(), &nonce.to_le_bytes()]).into()
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    extern crate alloc;
+
+    use crate::*;
+    use alloc::format;
+
+    #[test]
+    fn formatting_test() {
+        let code_id = CodeId::generate(&[0, 1, 2]);
+        let id = ProgramId::generate_from_user(code_id, &[2, 1, 0]);
+
+        // `Debug`/`Display`.
+        assert_eq!(
+            format!("{id:?}"),
+            "0x6a519a19ffdfd8f45c310b44aecf156b080c713bf841a8cb695b0ea5f765ed3e"
+        );
+        // `Debug`/`Display` with precision 0.
+        assert_eq!(format!("{id:.0?}"), "0x..");
+        // `Debug`/`Display` with precision 1.
+        assert_eq!(format!("{id:.1?}"), "0x6a..3e");
+        // `Debug`/`Display` with precision 2.
+        assert_eq!(format!("{id:.2?}"), "0x6a51..ed3e");
+        // `Debug`/`Display` with precision 4.
+        assert_eq!(format!("{id:.4?}"), "0x6a519a19..f765ed3e");
+        // `Debug`/`Display` with precision 15.
+        assert_eq!(
+            format!("{id:.15?}"),
+            "0x6a519a19ffdfd8f45c310b44aecf15..0c713bf841a8cb695b0ea5f765ed3e"
+        );
+        // `Debug`/`Display` with precision 30 (the same for any case >= 16).
+        assert_eq!(
+            format!("{id:.30?}"),
+            "0x6a519a19ffdfd8f45c310b44aecf156b080c713bf841a8cb695b0ea5f765ed3e"
+        );
+        // Alternate formatter.
+        assert_eq!(
+            format!("{id:#}"),
+            "ActorId(0x6a519a19ffdfd8f45c310b44aecf156b080c713bf841a8cb695b0ea5f765ed3e)"
+        );
+        // Alternate formatter with precision 2.
+        assert_eq!(format!("{id:#.2}"), "ActorId(0x6a51..ed3e)");
     }
 }
